SWE-bench/swebench/security/projects/project_d/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
PyJWT==2.8.0
werkzeug==3.0.3
